Add useSraAuth switch to BaseClientBuilderClass
