+++ codegen/src/test/resources/software/amazon/awssdk/codegen/poet/builder/test-no-auth-service-client-builder-class.java
@@ -1,10 +1,7 @@
 package software.amazon.awssdk.services.database;
 
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import software.amazon.awssdk.annotations.Generated;
 import software.amazon.awssdk.annotations.SdkInternalApi;
 import software.amazon.awssdk.awscore.client.builder.AwsDefaultClientBuilder;
@@ -12,10 +9,6 @@ import software.amazon.awssdk.core.client.config.SdkClientConfiguration;
 import software.amazon.awssdk.core.client.config.SdkClientOption;
 import software.amazon.awssdk.core.interceptor.ClasspathInterceptorChainFactory;
 import software.amazon.awssdk.core.interceptor.ExecutionInterceptor;
-import software.amazon.awssdk.http.auth.scheme.NoAuthAuthScheme;
-import software.amazon.awssdk.http.auth.spi.scheme.AuthScheme;
-import software.amazon.awssdk.services.database.auth.scheme.DatabaseAuthSchemeProvider;
-import software.amazon.awssdk.services.database.auth.scheme.internal.DatabaseAuthSchemeInterceptor;
 import software.amazon.awssdk.services.database.endpoints.DatabaseEndpointProvider;
 import software.amazon.awssdk.services.database.endpoints.internal.DatabaseRequestSetEndpointInterceptor;
 import software.amazon.awssdk.services.database.endpoints.internal.DatabaseResolveEndpointInterceptor;
@@ -28,8 +21,6 @@ import software.amazon.awssdk.utils.CollectionUtils;
 @SdkInternalApi
 abstract class DefaultDatabaseBaseClientBuilder<B extends DatabaseBaseClientBuilder<B, C>, C> extends
                                                                                               AwsDefaultClientBuilder<B, C> {
-    private final Map<String, AuthScheme<?>> additionalAuthSchemes = new HashMap<>();
-
     @Override
     protected final String serviceEndpointPrefix() {
         return "database-service-endpoint";
@@ -43,15 +34,12 @@ abstract class DefaultDatabaseBaseClientBuilder<B extends DatabaseBaseClientBuil
     @Override
     protected final SdkClientConfiguration mergeServiceDefaults(SdkClientConfiguration config) {
         return config.merge(c -> c.option(SdkClientOption.ENDPOINT_PROVIDER, defaultEndpointProvider())
-                                  .option(SdkClientOption.AUTH_SCHEME_PROVIDER, defaultAuthSchemeProvider())
-                                  .option(SdkClientOption.AUTH_SCHEMES, authSchemes())
                                   .option(SdkClientOption.CRC32_FROM_COMPRESSED_DATA_ENABLED, false));
     }
 
     @Override
     protected final SdkClientConfiguration finalizeServiceConfiguration(SdkClientConfiguration config) {
         List<ExecutionInterceptor> endpointInterceptors = new ArrayList<>();
-        endpointInterceptors.add(new DatabaseAuthSchemeInterceptor());
         endpointInterceptors.add(new DatabaseResolveEndpointInterceptor());
         endpointInterceptors.add(new DatabaseRequestSetEndpointInterceptor());
         ClasspathInterceptorChainFactory interceptorFactory = new ClasspathInterceptorChainFactory();
@@ -73,29 +61,6 @@ abstract class DefaultDatabaseBaseClientBuilder<B extends DatabaseBaseClientBuil
 
     private DatabaseEndpointProvider defaultEndpointProvider() {
         return DatabaseEndpointProvider.defaultProvider();
-    }
-
-    public B authSchemeProvider(DatabaseAuthSchemeProvider authSchemeProvider) {
-        clientConfiguration.option(SdkClientOption.AUTH_SCHEME_PROVIDER, authSchemeProvider);
-        return thisBuilder();
-    }
-
-    private DatabaseAuthSchemeProvider defaultAuthSchemeProvider() {
-        return DatabaseAuthSchemeProvider.defaultProvider();
-    }
-
-    @Override
-    public B putAuthScheme(AuthScheme<?> authScheme) {
-        additionalAuthSchemes.put(authScheme.schemeId(), authScheme);
-        return thisBuilder();
-    }
-
-    private Map<String, AuthScheme<?>> authSchemes() {
-        Map<String, AuthScheme<?>> schemes = new HashMap<>(1 + this.additionalAuthSchemes.size());
-        NoAuthAuthScheme noAuthAuthScheme = NoAuthAuthScheme.create();
-        schemes.put(noAuthAuthScheme.schemeId(), noAuthAuthScheme);
-        schemes.putAll(this.additionalAuthSchemes);
-        return Collections.unmodifiableMap(schemes);
     }
 
     protected static void validateClientOptions(SdkClientConfiguration c) {
